Correct various bugs
1) Bad specification in sharedGoalB.xml which crashes in debug. Ill-formed
   AABB where min point was greater than max point.
2) Fix the set_xproperty method. Entering the action would optionally set
   state based on _undoOnExit. But leaving would search for that state
   regardless. Now leaving is likewise dependent on _undoOnExit.
3) Nav mesh had a 64-bit error. I was comparing a 32-bit unsigned int with
   a 64-bit unsigned -1 value. No longer equal. Changed (to a very limited
   degree) unsigned int to size_t.

+++ src/Menge/MengeCore/resources/NavMesh.cpp
@@ -301,17 +301,15 @@ namespace Menge {
 		}
 
 		std::vector< bool > processed( _obstCount, false );
-		for ( unsigned int o = 0; o < (unsigned int)_obstCount; ++o ) {
+		for ( size_t o = 0; o < _obstCount; ++o ) {
 			_obstacles[ o ]._id = o;
 			if ( processed[ o ] ) continue;
-			const unsigned int START = o;
-			unsigned int curr = o;
+			const size_t START = o;
+			size_t curr = o;
 			while ( curr != NavMeshObstacle::NO_NEIGHBOR_OBST && !processed[ curr ] ) {
 				processed[ curr ] = true;
 				NavMeshObstacle & obst = _obstacles[ curr ];
 				size_t nID = reinterpret_cast<size_t>( obst._node );
-                //error: cast from 'Menge::Agents::Obstacle*' to 'unsigned int' loses precision
-				// [-fpermissive]
 				assert( nID < _nCount && "Finalizing invalid node id from obstacle" );
 				obst._node = &_nodes[ nID ];
